shot-evaluation/requirements.txt
azure-core==1.31.0
azure-storage-blob==12.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.9.0
greenlet==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
